format

+++ src/user/user.service.ts
@@ -1,4 +1,8 @@
-import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectModel } from '@nestjs/mongoose';
@@ -8,7 +12,7 @@ import { Model } from 'mongoose';
 
 @Injectable()
 export class UserService {
-  constructor(@InjectModel(Users.name) private userSchema: Model<Users>) { }
+  constructor(@InjectModel(Users.name) private userSchema: Model<Users>) {}
 
   async create(createUserDto: CreateUserDto) {
     const encryptedPass = await bcrypt.hash(
@@ -35,15 +39,13 @@ export class UserService {
     if (user) {
       if (user._id.toString() === userId)
         return await this.userSchema.findByIdAndUpdate(_id, updateUserDto);
-      else throw new ForbiddenException("Can not mutate other entity.")
-    }
-    else throw new NotFoundException('user not found.');
+      else throw new ForbiddenException('Can not mutate other entity.');
+    } else throw new NotFoundException('user not found.');
   }
 
   async remove(_id: string) {
     const user = await this.userSchema.findById(_id);
-    if (user)
-      return await this.userSchema.findByIdAndDelete({ _id });
+    if (user) return await this.userSchema.findByIdAndDelete({ _id });
     else throw new NotFoundException('user not found.');
   }
